+++ src/grepros/main.py
@@ -118,30 +118,30 @@ Export all bag messages to SQLite and Postgres, print only export progress:
                   "Keyword arguments are given to output writer."),
 
         dict(args=["--plugin"],
-             dest="PLUGINS", metavar="PLUGIN", nargs="+", default=[],
+             dest="PLUGINS", metavar="PLUGIN", nargs="+", default=[], action="append",
              help="load a Python module or class as plugin"),
     ],
 
     "groups": {"Filtering": [
 
         dict(args=["-t", "--topic"],
-             dest="TOPICS", metavar="TOPIC", nargs="+", default=[],
+             dest="TOPICS", metavar="TOPIC", nargs="+", default=[], action="append",
              help="ROS topics to scan if not all (supports * wildcards)"),
 
         dict(args=["-nt", "--no-topic"],
-             dest="SKIP_TOPICS", metavar="TOPIC", nargs="+", default=[],
+             dest="SKIP_TOPICS", metavar="TOPIC", nargs="+", default=[], action="append",
              help="ROS topics to skip (supports * wildcards)"),
 
         dict(args=["-d", "--type"],
-             dest="TYPES", metavar="TYPE", nargs="+", default=[],
+             dest="TYPES", metavar="TYPE", nargs="+", default=[], action="append",
              help="ROS message types to scan if not all (supports * wildcards)"),
 
         dict(args=["-nd", "--no-type"],
-             dest="SKIP_TYPES", metavar="TYPE", nargs="+", default=[],
+             dest="SKIP_TYPES", metavar="TYPE", nargs="+", default=[], action="append",
              help="ROS message types to skip (supports * wildcards)"),
 
         dict(args=["--condition"],
-             dest="CONDITIONS", metavar="CONDITION", nargs="+", default=[],
+             dest="CONDITIONS", metavar="CONDITION", nargs="+", default=[], action="append",
              help="extra conditions to require for matching messages,\n"
                   "as ordinary Python expressions, can refer to last messages\n"
                   "in topics as {topic /my/topic}; topic name can contain wildcards.\n"
@@ -191,12 +191,12 @@ Export all bag messages to SQLite and Postgres, print only export progress:
              help="emit every Nth match in topic"),
 
         dict(args=["-sf", "--select-field"],
-             dest="SELECT_FIELDS", metavar="FIELD", nargs="*", default=[],
+             dest="SELECT_FIELDS", metavar="FIELD", nargs="+", default=[], action="append",
              help="message fields to use in matching if not all\n"
                   "(supports nested.paths and * wildcards)"),
 
         dict(args=["-ns", "--no-select-field"],
-             dest="NOSELECT_FIELDS", metavar="FIELD", nargs="*", default=[],
+             dest="NOSELECT_FIELDS", metavar="FIELD", nargs="+", default=[], action="append",
              help="message fields to skip in matching\n"
                   "(supports nested.paths and * wildcards)"),
 
@@ -235,12 +235,12 @@ Export all bag messages to SQLite and Postgres, print only export progress:
                   "around match"),
 
         dict(args=["-pf", "--print-field"],
-             dest="PRINT_FIELDS", metavar="FIELD", nargs="*", default=[],
+             dest="PRINT_FIELDS", metavar="FIELD", nargs="+", default=[], action="append",
              help="message fields to print in console output if not all\n"
                   "(supports nested.paths and * wildcards)"),
 
         dict(args=["-np", "--no-print-field"],
-             dest="NOPRINT_FIELDS", metavar="FIELD", nargs="*", default=[],
+             dest="NOPRINT_FIELDS", metavar="FIELD", nargs="+", default=[], action="append",
              help="message fields to skip in console output\n"
                   "(supports nested.paths and * wildcards)"),
 
@@ -310,12 +310,12 @@ Export all bag messages to SQLite and Postgres, print only export progress:
     ], "Bag input control": [
 
         dict(args=["-n", "--filename"],
-             dest="FILES", metavar="FILE", nargs="*", default=[],
+             dest="FILES", metavar="FILE", nargs="+", default=[], action="append",
              help="names of ROS bagfiles to scan if not all in directory\n"
                   "(supports * wildcards)"),
 
         dict(args=["-p", "--path"],
-             dest="PATHS", metavar="PATH", nargs="*", default=[],
+             dest="PATHS", metavar="PATH", nargs="+", default=[], action="append",
              help="paths to scan if not current directory\n"
                   "(supports * wildcards)"),
 
@@ -388,25 +388,14 @@ def make_parser():
     return argparser
 
 
-def validate_args(args):
+def process_args(args):
     """
-    Validates arguments, prints errors, returns success.
+    Converts or combines arguments where necessary, returns args.
 
     @param   args  arguments object like argparse.Namespace
     """
-    errors = collections.defaultdict(list)  # {category: [error, ]}
     if args.CONTEXT:
         args.BEFORE = args.AFTER = args.CONTEXT
-
-    # Validate --write .. key=value
-    for opts in args.DUMP_TARGET:  # List of lists, one for each --write
-        erropts = []
-        for opt in opts[1:]:
-            try: dict([opt.split("=", 1)])
-            except Exception: erropts.append(opt)
-        if erropts:
-            errors[""].append('Invalid KEY=VALUE in "--write %s": %s' %
-                              (" ".join(opts), " ".join(erropts)))
 
     # Default to printing metadata for publish/write if no console output
     args.VERBOSE = False if args.SKIP_VERBOSE else (args.VERBOSE or not args.CONSOLE)
@@ -418,9 +407,39 @@ def validate_args(args):
     args.LINE_PREFIX = args.LINE_PREFIX and (args.RECURSE or len(args.FILES) != 1
                                              or args.PATHS or any("*" in x for x in args.FILES))
 
-    for k, v in vars(args).items():  # Drop duplicates from list values
+    for k, v in vars(args).items():  # Flatten lists  of lists and drop duplicates
         if k != "DUMP_TARGET" and isinstance(v, list):
-            setattr(args, k, list(collections.OrderedDict((x, None) for x in v)))
+            here = set()
+            setattr(args, k, [x for xx in v for x in (xx if isinstance(xx, list) else [xx])
+                              if not (x in here or here.add(x))])
+
+    for n, v in [("START_TIME", args.START_TIME), ("END_TIME", args.END_TIME)]:
+        if v is None: continue  # for v, n
+        try: v = float(v)
+        except Exception: pass
+        try: not isinstance(v, float) and setattr(args, n, parse_datetime(v))
+        except Exception: pass
+
+    return  args
+
+
+def validate_args(args):
+    """
+    Validates arguments, prints errors, returns success.
+
+    @param   args  arguments object like argparse.Namespace
+    """
+    errors = collections.defaultdict(list)  # {category: [error, ]}
+
+    # Validate --write .. key=value
+    for opts in args.DUMP_TARGET:  # List of lists, one for each --write
+        erropts = []
+        for opt in opts[1:]:
+            try: dict([opt.split("=", 1)])
+            except Exception: erropts.append(opt)
+        if erropts:
+            errors[""].append('Invalid KEY=VALUE in "--write %s": %s' %
+                              (" ".join(opts), " ".join(erropts)))
 
     for n, v in [("START_TIME", args.START_TIME), ("END_TIME", args.END_TIME)]:
         if v is None: continue  # for v, n
@@ -465,7 +484,7 @@ def flush_stdout():
 def preload_plugins():
     """Imports and initializes plugins from auto-load folder and from arguments."""
     args = make_parser().parse_known_args()[0] if "--plugin" in sys.argv else None
-    try: plugins.init(args)
+    try: plugins.init(process_args(args) if args else None)
     except ImportWarning: sys.exit(1)
 
 
@@ -490,7 +509,7 @@ def run():
     source, sink = None, None
     try:
         ConsolePrinter.configure({"always": True, "never": False}.get(args.COLOR))
-        if not validate_args(args):
+        if not validate_args(process_args(args)):
             sys.exit(1)
 
         source = plugins.load("source", args) or \